Added confirm tour/stop delete

+++ js/components/StopBox.vue
@@ -1,15 +1,23 @@
 <template>
-    <div class="stop-box bg-fit" :style="backgroundImage" @click="$emit('click')">
-        <div class="white-circle">{{ stop.order }}</div>
+    <div>
+        <div class="stop-box bg-fit" :style="backgroundImage" @click="$emit('click')">
+            <div class="white-circle">{{ stop.order }}</div>
 
-        <div class="title">
-            {{ stop.title }}
+            <div class="title">
+                {{ stop.title }}
+            </div>
+
+            <div class="delete" @click.stop="deleteStop">
+                <fa v-if="busy" class="fa-spin" :icon="['fas', 'spinner']" />
+                <fa v-else :icon="['fas', 'times']" />
+            </div>
+
         </div>
 
-        <div class="delete" @click.stop="deleteStop">
-            <fa v-if="busy" class="fa-spin" :icon="['fas', 'spinner']" />
-            <fa v-else :icon="['fas', 'times']" />
-        </div>
+        <!-- Confirmation modal -->
+        <confirm-modal ref="confirm" yesButton="Delete">
+            Are you sure you want to delete this stop?
+        </confirm-modal>
     </div>
 </template>
 
@@ -22,6 +30,7 @@ export default {
 
     data: () => ({
         busy: false,
+        confirmModal: false,
     }),
 
     computed: {
@@ -40,17 +49,19 @@ export default {
     },
 
     methods: {
-        deleteStop() {
-            this.busy = true;
-            axios.delete(this.saveTourUrl + `/stops/${this.stop.id}`)
-                .then( ({ data }) => {
-                    this.$emit('deleted');
-                    this.busy = false;
-                })
-                .catch( e => {
-                    alerts.addMessage('error', e.response.data.message);
-                    this.busy = false;
-                })
+        deleteStop(confirm = true) {
+            this.$refs.confirm.confirm(() => {
+                this.busy = true;
+                axios.delete(this.saveTourUrl + `/stops/${this.stop.id}`)
+                    .then( ({ data }) => {
+                        this.$emit('deleted');
+                        this.busy = false;
+                    })
+                    .catch( e => {
+                        alerts.addMessage('error', e.response.data.message);
+                        this.busy = false;
+                    })
+            });
         },
     },
 }
